Move stitching tools out of control flow
